Move SnapshotStore to more general logic package
